Add note about miner's interaction with level GUI
This is just a message to clarify the behavior of the miner's interaction with this GUI. I was unable to find an explanation for why the miner was ignoring my command to mine at a specified level and instead kept leveling the miner's hut further increasing his workload for building the shaft. I finally read through the source code just to figure it out. I would really recommend making a change to the GUI to reflect that no action will be taken until the mine shaft is completed. That level of feedback would be welcome to newcomers like me who believed the miner's failure to produce resources ruined the entire mod's balance. This message here is just a quick solution in my opinion.