updated manual modpack install section (#632)
* updated manual modpack install section

updates to manual install section to remove outdated info as curseforge no longer distributes mods with the pack download due to licensing issues. 
While still technically possible to manually download, doing so is so complex it is probably outside scope of the wiki. the included options and explanation should be enough for most users.

* Title change

* Update source/installation/java.md

* Update source/installation/java.md

* Update source/installation/java.md

Co-authored-by: ravenbuilder934 <sabrinajamie@outlook.com>