Add DO shingles info to shingles page (#701)
* Add files via upload

Just two images for me to add DO info on shingles page

* Delete cutter.PNG

* Delete mouseover.PNG

* Add files via upload

two images for updated shingles page

* Delete cutter.png

* Delete mouseover.png

* Add files via upload

* Update shingles.md

Ok! Let's see if I got it right this time?

Hopefully this will add some info for DO shingles at the bottom of the page, and HOPEFULLY the text at the top will link to the bottom!