Update stash.md
Add that the stash block can accept hoppers and be used to automate pick up of items created outside of the Minecolonies mod such as mob farms, vine farms, and with machines created in other mods such as Create or Immersive Engineering.  Many may not know that this block can be used in this way to automate and that this brings alot of cross compatibility between other mods and Minecolonies.