Update nethermine.md
Added a Note stating that this building is not yet available in all colony styles. This was a point of confusion for myself and others I have seen online and I feel is worth noting on the wiki. Obviously this can change in future so I added the version number as reference.