Add Structurize Frame crafting info to Carpenter
As JEI shows the recipe for Structurize blocks not needing the Build Tool I was a bit stumped on what if even I had to teach the Carpenter for them to craft those (and didn't just try but searched a while until I found some post on Reddit about it). I'd say it would be good to have this info on the wiki page directly, since this was the first place I looked, but couldn't find anything.